xcm config changes
